devide fat single file
