+++ frontend/src/pages/AccountPage/AccountNotFound.tsx
@@ -1,0 +1,9 @@
+import { VFC } from 'react';
+
+export const AccountNotFound: VFC = () => {
+  return (
+    <div>
+      <p>アカウントが見つかりませんでした</p>
+    </div>
+  );
+};
